+++ src/utils/VariantType.hpp
@@ -34,14 +34,15 @@ namespace luagdextension {
  * Object that represents Godot's builtin classes (a.k.a. Variants) in Lua.
  */
 class VariantType {
-protected:
-	Variant::Type type;
-
 public:
 	VariantType(Variant::Type type);
+	static VariantType from_variant(const Variant& value);
 
 	Variant::Type get_type() const;
-	String get_type_name() const;
+	bool has_type_hints() const;
+	PropertyHint get_property_hint() const;
+	String get_property_hint_string() const;
+	String to_string() const;
 
 	Variant construct_default() const;
 	Variant construct(const sol::variadic_args& args) const;
@@ -49,6 +50,25 @@ public:
 	bool operator==(const VariantType& other) const;
 
 	static void register_usertype(sol::state_view& state);
+
+protected:
+	Variant::Type type;
+	
+	// Used for typed arrays and typed dictionaries
+	Variant subtype1;
+	// Used for typed dictionaries
+	Variant subtype2;
+
+	VariantType(Variant::Type type, const Variant& subtype1, const Variant& subtype2);
+
+	static sol::object __index(sol::this_state L, const VariantType& cls, const sol::stack_object& key);
+	
+	static std::tuple<Variant::Type, StringName, Variant> subtype_to_constructor_args(const Variant& subtype);
+	static String subtype_name(const Variant& subtype);
+	static String subtype_hint_string(const Variant& subtype);
+	static Variant get_subtype(const Array& array);
+	static Variant get_key_subtype(const Dictionary& dict);
+	static Variant get_value_subtype(const Dictionary& dict);
 };
 
 }